+++ upgrades/0.0.99.seperate-ns-from-app/pkg/argocd/application.go
@@ -7,7 +7,7 @@ import (
 	"io"
 	"text/template"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.97.seperate-ns-from-app/pkg/lib/manifest/apis/okctl.io/v1alpha1"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.99.seperate-ns-from-app/pkg/lib/manifest/apis/okctl.io/v1alpha1"
 )
 
 const (
