+++ upgrades/0.0.99.seperate-ns-from-app/pkg/argocd/templates/namespaces-readme.md
