+++ upgrades/0.0.99.seperate-ns-from-app/pkg/argocd/testdata/argoapp-Should setup correct files and directories.golden
