+++ upgrades/0.0.99.seperate-ns-from-app/pkg/paths/api.go
@@ -6,7 +6,7 @@ import (
 	"os/exec"
 	"path"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.97.seperate-ns-from-app/pkg/lib/manifest/apis/okctl.io/v1alpha1"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.99.seperate-ns-from-app/pkg/lib/manifest/apis/okctl.io/v1alpha1"
 )
 
 const (
